refactor(session): replace legacy Session and SessionRepository classes
