+++ tomcat/src/main/java/nextstep/jwp/controller/AuthController.java
@@ -6,24 +6,21 @@ import nextstep.jwp.servlet.handler.RequestMapping;
 import org.apache.coyote.servlet.cookie.HttpCookie;
 import org.apache.coyote.servlet.request.HttpRequest;
 import org.apache.coyote.servlet.response.ResponseEntity;
-import org.apache.coyote.servlet.session.Session2;
-import org.apache.coyote.servlet.session.SessionRepository;
+import org.apache.coyote.servlet.session.Session;
 import org.apache.coyote.support.HttpMethod;
 
 public class AuthController {
 
     private final UserService userService;
-    private final SessionRepository sessionRepository;
 
-    public AuthController(UserService userService, SessionRepository sessionRepository) {
+    public AuthController(UserService userService) {
         this.userService = userService;
-        this.sessionRepository = sessionRepository;
     }
 
     @RequestMapping(method = HttpMethod.GET, path = "/login")
     public ResponseEntity getLoginPage(HttpRequest request) {
         final var session = request.getSession();
-        if (session.hasAttribute(Session2.USER_ATTRIBUTE)) {
+        if (session.hasAttribute(Session.USER_ATTRIBUTE)) {
             return ResponseEntity.redirect("/index.html").build();
         }
         return ResponseEntity.ok().setViewResource("/login.html").build();
@@ -33,7 +30,7 @@ public class AuthController {
     public ResponseEntity login(HttpRequest request) {
         final var user = userService.login(DtoAssembler.ofLoginDto(request));
         final var session = request.getSession();
-        session.setAttribute(Session2.USER_ATTRIBUTE, user);
+        session.setAttribute(Session.USER_ATTRIBUTE, user);
         final var sessionCookie = HttpCookie.ofSession(session);
         return ResponseEntity.redirect("/index.html")
                 .setCookie(sessionCookie)
