+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -1,15 +1,14 @@
 package org.apache.catalina.connector;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.net.ServerSocket;
+import java.net.Socket;
 import org.apache.coyote.http11.Http11Processor;
 import org.apache.coyote.servlet.Servlet;
 import org.apache.coyote.servlet.session.SessionRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.net.ServerSocket;
-import java.net.Socket;
 
 public class Connector implements Runnable {
