+++ tomcat/src/main/java/org/apache/coyote/servlet/cookie/HttpCookie.java
@@ -1,7 +1,6 @@
 package org.apache.coyote.servlet.cookie;
 
 import org.apache.coyote.servlet.session.Session;
-import org.apache.coyote.servlet.session.Session2;
 import org.apache.coyote.support.HttpException;
 
 public class HttpCookie {
@@ -32,7 +31,7 @@ public class HttpCookie {
         return new HttpCookie(keyValues[0], keyValues[1]);
     }
 
-    public static HttpCookie ofSession(Session2 session) {
+    public static HttpCookie ofSession(Session session) {
         return new HttpCookie(Session.JSESSIONID, session.getId(), Session.VALIDITY_IN_SECONDS);
     }
