+++ tomcat/src/main/java/org/apache/coyote/servlet/request/RequestHeaders.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Map;
 import org.apache.coyote.servlet.cookie.HttpCookie;
 import org.apache.coyote.servlet.cookie.HttpCookies;
-import org.apache.coyote.servlet.session.Session2;
+import org.apache.coyote.servlet.session.Session;
 import org.apache.coyote.support.HttpException;
 
 public class RequestHeaders {
@@ -54,7 +54,7 @@ public class RequestHeaders {
 
     public HttpCookie getSessionCookie() {
         final var cookies = getCookies();
-        return cookies.getCookie(Session2.JSESSIONID);
+        return cookies.getCookie(Session.JSESSIONID);
     }
 
     public HttpCookies getCookies() {
